requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
newsapi-python==0.2.7
psycopg2==2.9.9
psycopg2-binary==2.9.9
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
